Switch time tracking from whole days to fractional (hour-precision) days
El sistema solo medía el tiempo en dias in-game ENTEROS
(TimeProvider#getCurrentDay devolvia long), asi que la pudricion
quedaba congelada todo el dia y saltaba de golpe recien al cruzar la
medianoche in-game — nada progresivo, y particularmente malo dado
que RealisticSeasons permite alargar/acortar la duracion real de un
dia in-game por configuracion.

TimeProvider#getCurrentDay ahora devuelve double (ej. 572.34),
combinando el dia calendario (Calendar#getTotalDays) con la hora del
dia actual (SeasonsAPI#getHours/getMinutes/getSeconds). Esto se
propaga a traves de FoodManager, CatchUpProcessor, ApplianceGUI,
InventoryListener, ConsumeListener, FarmingListener, WateringManager
y RSDebugCommand (todos los watermarks de PDC pasan de LONG a DOUBLE).

Los recursos discretos (Hielo, cargas de agua del Aspersor) no pueden
consumirse en fracciones: el consumo redondea hacia arriba (ceil) el
tramo frio cubierto, priorizando nunca dejar un periodo parcial sin
proteger por sobre un ahorro marginal de combustible.

+++ src/main/java/cl/nico/realisticsurvival/api/time/TimeProvider.java
@@ -18,14 +18,22 @@ import org.bukkit.entity.Player;
 public interface TimeProvider {
 
     /**
-     * Dia in-game absoluto del mundo indicado. Es la unidad base de tiempo de todo el
-     * plugin: la pudricion de alimentos, el hielo de los electrodomesticos y la humedad
-     * del farmland se calculan siempre como diferencia entre dos "dias absolutos".
+     * Dia in-game absoluto del mundo indicado, CON precision de hora — ej. {@code 572.34}
+     * es aproximadamente las 8:10am del dia 572 (la parte decimal es la fraccion del dia
+     * transcurrida). Es la unidad base de tiempo de todo el plugin: la pudricion de
+     * alimentos, el hielo de los electrodomesticos y la humedad del farmland se calculan
+     * siempre como diferencia entre dos de estos valores.
+     * <p>
+     * Se usa precision de hora (no solo el dia calendario entero) para que la pudricion
+     * avance de forma continua durante el dia, en vez de saltar solo al cruzar la
+     * medianoche — importante porque RealisticSeasons permite alargar/acortar la duracion
+     * real de un dia in-game por configuracion, asi que "1 dia" no es una unidad de tiempo
+     * real fija.
      *
      * @param world mundo a consultar
-     * @return dia in-game absoluto (nunca negativo, monotonamente creciente)
+     * @return dia in-game absoluto con fraccion horaria (nunca negativo, monotonamente creciente)
      */
-    long getCurrentDay(World world);
+    double getCurrentDay(World world);
 
     /**
      * Temperatura ambiente (en grados) de una ubicacion, segun RealisticSeasons
